+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Most recent releases are shown at the top. Each release shows:
 ### fixed:
 - move query preprocessing to base classes (#209)
 - fixed extra kwargs handling for LangChain LLMs (#186)
+- Make keyword fields in WhooshStore case insensitive. (#211)
+- `doc_from_dict` is modifying the original dictionary (#211)
 
 
 ## 0.16.3 (2025-07-20)