+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
-## 0.16.4 (TBD)
+## 0.17.0 (TBD)
 
 ### new:
 - Support for RAG with SharePoint (#210)